little arch fixes
