+++ app/src/main/java/com/example/playlistmaker/di/RepositoryModule.kt
@@ -1,13 +1,11 @@
 package com.example.playlistmaker.di
 
 import com.example.playlistmaker.data.player.repository.AndroidMediaPlayerRepositoryImpl
-import com.example.playlistmaker.data.playlist.repository.PlaylistCoverRepositoryImpl
 import com.example.playlistmaker.data.playlist.repository.PlaylistRepositoryImpl
 import com.example.playlistmaker.data.search.repository.SearchHistoryRepositoryImpl
 import com.example.playlistmaker.data.search.repository.TrackRepositoryImpl
 import com.example.playlistmaker.data.settings.repository.SettingsRepositoryImpl
 import com.example.playlistmaker.domain.player.api.repository.MediaPlayerRepository
-import com.example.playlistmaker.domain.playlist.api.repository.PlaylistCoverRepository
 import com.example.playlistmaker.domain.playlist.api.repository.PlaylistRepository
 import com.example.playlistmaker.domain.search.api.repository.SearchHistoryRepository
 import com.example.playlistmaker.domain.search.api.repository.TrackRepository
@@ -34,10 +32,6 @@ val repositoryModule = module {
     }
 
     single<PlaylistRepository> {
-        PlaylistRepositoryImpl(get())
-    }
-
-    single<PlaylistCoverRepository> {
-        PlaylistCoverRepositoryImpl(get())
+        PlaylistRepositoryImpl(get(), get())
     }
 }
