+++ app/src/main/java/com/example/playlistmaker/di/ViewModelModule.kt
@@ -35,11 +35,11 @@ val viewModelModule = module {
     }
 
     viewModel {
-        NewPlaylistViewModel(get(), get())
+        NewPlaylistViewModel(get())
     }
 
     viewModel {
-        PlaylistsViewModel(get(), get())
+        PlaylistsViewModel(get())
     }
 
 }
