--- app/src/main/java/com/example/playlistmaker/domain/playlist/api/repository/PlaylistCoverRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.playlist.api.repository
-
-interface PlaylistCoverRepository {
-    fun getImageFromPrivateStorage(fileName: String): String
-
-    fun saveImageToPrivateStorage(uri: String, fileName: String)
-}
