+++ app/src/main/java/com/example/playlistmaker/domain/playlist/api/repository/PlaylistRepository.kt
@@ -6,4 +6,8 @@ interface PlaylistRepository {
     suspend fun savePlaylist(playlist: Playlist)
 
     suspend fun getPlaylists(): List<Playlist>
+
+    fun getImageFromPrivateStorage(fileName: String): String
+
+    fun saveImageToPrivateStorage(uri: String, fileName: String)
 }
