+++ app/src/main/java/com/example/playlistmaker/ui/new_playlist/view_model/NewPlaylistViewModel.kt
@@ -4,12 +4,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.playlistmaker.domain.playlist.api.repository.PlaylistRepository
 import com.example.playlistmaker.domain.model.Playlist
-import com.example.playlistmaker.domain.playlist.api.repository.PlaylistCoverRepository
 import kotlinx.coroutines.launch
 
 class NewPlaylistViewModel(
     private val playlistRepository: PlaylistRepository,
-    private val playlistCoverRepository: PlaylistCoverRepository
 ) : ViewModel() {
     fun createPlaylist(playlist: Playlist) {
         viewModelScope.launch {
@@ -18,6 +16,6 @@ class NewPlaylistViewModel(
     }
 
     fun saveImage(uri: String, fileName: String) {
-        playlistCoverRepository.saveImageToPrivateStorage(uri, fileName)
+        playlistRepository.saveImageToPrivateStorage(uri, fileName)
     }
 }
