+++ app/src/main/java/com/example/playlistmaker/ui/playlists/recycler/PlaylistViewHolder.kt
@@ -1,22 +1,17 @@
 package com.example.playlistmaker.ui.playlists.recycler
 
-import android.net.Uri
-import android.os.Environment
-import android.util.Log
 import androidx.core.net.toUri
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.R
 import com.example.playlistmaker.databinding.PlaylistViewBinding
 import com.example.playlistmaker.domain.model.Playlist
-import com.example.playlistmaker.domain.playlist.api.repository.PlaylistCoverRepository
-import com.example.playlistmaker.ui.util.Convert
+import com.example.playlistmaker.domain.playlist.api.repository.PlaylistRepository
 import java.io.File
 
 class PlaylistViewHolder(
     private val binding: PlaylistViewBinding,
-    private val playlistCoverRepository: PlaylistCoverRepository,
+    private val playlistRepository: PlaylistRepository,
 ) :
     RecyclerView.ViewHolder(binding.root) {
 
@@ -25,7 +20,7 @@ class PlaylistViewHolder(
         //TODO сделать правильные склонения имен
         tvTracksCount.text = "${model.tracksCount} треков"
         Glide.with(itemView)
-            .load(File(playlistCoverRepository.getImageFromPrivateStorage(model.name).toUri().path))
+            .load(File(playlistRepository.getImageFromPrivateStorage(model.name).toUri().path))
             .placeholder(R.drawable.track_placeholder)
             .centerCrop()
             .into(picture)
