+++ app/src/main/java/com/example/playlistmaker/ui/playlists/view_model/PlaylistsViewModel.kt
@@ -4,14 +4,12 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.playlistmaker.domain.playlist.api.repository.PlaylistCoverRepository
 import com.example.playlistmaker.domain.playlist.api.repository.PlaylistRepository
 import com.example.playlistmaker.ui.playlists.model.PlaylistState
 import kotlinx.coroutines.launch
 
 class PlaylistsViewModel(
     private val playlistRepository: PlaylistRepository,
-    private val playlistCoverRepository: PlaylistCoverRepository,
 ) : ViewModel() {
 
     private val stateLiveData = MutableLiveData<PlaylistState>()
@@ -28,7 +26,7 @@ class PlaylistsViewModel(
         }
     }
 
-    fun getCoverRepository(): PlaylistCoverRepository {
-        return playlistCoverRepository
+    fun getCoverRepository(): PlaylistRepository {
+        return playlistRepository
     }
 }
